Add files via upload
The v2 version fixes a problem discovered by a user (Haim) of Wolf's original firmware when measuring frequencies in the range 1 .. 4 MHz. It is also present in my changed firmware. Below 4MHz is the highest frequency range the PIC measures without pre-scaler. If you measure signals with a duty cycle significantly different from 50% in that range, the pulse width can drop below the 120ns required by the PIC. This causes the PIC to show fluctuating values and it is quite obvious when it happens. It is more likely to happen if no pre-amp is used but can happen even with a pre-amp.  The fix is to use the pre-scaler earlier (above 983KHz) because that reduces the required pulse-width significantly. This reduces the measurement resolution from 4Hz to 8Hz in the range from 1..4MHz but is no issue because we can anyway only see the top 5 digits (display resolution 100Hz)
This version also fixes a bug in the timing loop which causes it to run a tiny bit too fast (I forgot to insert a NOP instruction), so even if you don't experience problems measuring in the 1-4MHz range you should change to this version.